remove resolvedRows from finished nodes
